+++ docs/changelog.md
@@ -2,7 +2,11 @@
 
 # Changelog
 
-# 2026-09-24
+## 2026-09-25
+
+- Fix Discourse database and backup connections when the database relation uses a non-default port.
+
+## 2026-09-24
 
 - chore(docs): Manually update the documentation files, and add a new
   workflow to automatically manage updates.